Another approach at replacing the doctype.
It appears that we can loadHTML the <!DOCTYPE and then use name, publicId and systemId to populate the DOMImplementation->createDocumentType().
This doctype can be used to replaceChild of the current document's doctype.
Returning saveXML formats the html5 empty doctype better and it is much less nurve wrecking to preg_replace the <?xml version="1.0" encoding="iso-8859-1" standalone="yes"?> than a doctype.
Both this and the previous commit solves the problem, I might lean towards this option more.