Allow Pseudo Selectors
I think that cover pseudo selectors in css template engine Respect is
very handy.
To do this, I created 2 methods in the class `Query` that comes if there
is a pseudo selector in
css selector and passes to a factory that will return the instance of a
pseudo selector and will
call the `apply` to "filter" the dial, leaving compatible with xpath.

Only did the 'first-child', do not know if this is the best way for such
a task.
Hopefully opinions and suggestions.